fix: app test params type

+++ app/src/main/kotlin/io/mysocialapp/app/customfield/controllers/MyController.kt
@@ -17,7 +17,7 @@ open class MyController {
         get() = QueryParameters().apply {
             req?.getParameter("usage_key")?.let { usageKey = it }
             req?.getParameter("parent_type")?.let { parentType = it }
-            req?.getParameter("parent_id")?.let { parentId = it }
+            req?.getParameter("parent_id")?.let { parentId = it.toLong() }
         }
 
 }
